added generalized parsing scheme to ADF (#24)
