+++ qmworks/packages/packages.py
@@ -77,6 +77,19 @@ class Result:
             "archive": self.archive,
             "project_name": self.project_name}
 
+    def __getattr__(self, prop):
+        """Returns a section of the results.
+
+        Example:
+
+        ..
+            dipole = result.dipole
+        """
+        if prop in self.prop_dict:
+            return self.get_property(prop)
+        else:
+            raise KeyError("Generic property '" + str(prop) + "' not defined")
+
     def get_property(self, prop):
         """
         Look for the optional arguments to parse a property, which are stored
