+++ qmworks/parsers/__init__.py
@@ -1,3 +1,4 @@
+from .adf_parser import kfreader
 from .cp2KParser import (readCp2KBasis, readCp2KCoeff, readCp2KOverlap,
                          read_cp2k_number_of_orbitals)
 from .gamess_parser import *
